Final version - delivered

+++ Projeto 2/client/api.c
@@ -8,6 +8,7 @@
 #include <string.h>
 #include "api.h"
 #include "common/io.h"
+#include "common/constants.h"
 
 struct Client* client = NULL; // Cliente
 
@@ -26,7 +27,7 @@ int ems_setup(char const* req_pipe_path, char const* resp_pipe_path, char const*
   strcpy(client->req_pipe_path, req_pipe_path);
   strcpy(client->resp_pipe_path, resp_pipe_path);
 
-  char message[82]; // 1 + 40 + 40 + 1
+  char message[MAX_SIZE_PATHS]; // 1 + 40 + 40 + 1
   message[0] = '1';
   strncpy(&message[1], req_pipe_path, 40);
   strncpy(&message[41], resp_pipe_path, 40);
@@ -37,9 +38,14 @@ int ems_setup(char const* req_pipe_path, char const* resp_pipe_path, char const*
     fprintf(stderr, "[ERR]: open server pipe failed: %s\n", strerror(errno));
     return 1;
   }
-  print_str_size(server_fd, message, 82);
-  close(server_fd);
-
+  if (print_str_size(server_fd, message, 82)) {
+    fprintf(stderr, "Error writing to server pipe\n");
+    return 1;
+  }
+  if(close(server_fd) == -1) {
+    fprintf(stderr, "Error closing server pipe\n");
+    return 1;
+  }
 
   // Abrir pipe de response para leitura
   int resp_fd = open(resp_pipe_path, O_RDONLY);
@@ -47,18 +53,23 @@ int ems_setup(char const* req_pipe_path, char const* resp_pipe_path, char const*
     fprintf(stderr, "[ERR]: open response pipe failed: %s\n", strerror(errno));
     return 1;
   }
-
   char response_setup[sizeof(int)];
 
   // Ler o session_id do response pipe
   ssize_t bytesRead = read(resp_fd, response_setup, sizeof(int));
   if (bytesRead == -1) {
     fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
-    close(resp_fd);
+    if (close(resp_fd) == -1) {
+      fprintf(stderr, "Error closing response pipe\n");
+      return 1;
+    }
     ems_quit();
     return 1;
   }
-  close(resp_fd);
+  if (close(resp_fd) == -1) {
+    fprintf(stderr, "Error closing response pipe\n");
+    return 1;
+  }
 
   // Associar session id ao named pipe do server
   int session_id;
@@ -79,8 +90,14 @@ int ems_quit(void) {
     fprintf(stderr, "[ERR]: open server pipe failed: %s\n", strerror(errno));
     return 1;
   }
-  print_str_size(req_fd, message, 1);
-  close(req_fd);
+  if (print_str_size(req_fd, message, 1)) {
+    fprintf(stderr, "Error writing to request pipe\n");
+    return 1;
+  }
+  if(close(req_fd) == -1) {
+    fprintf(stderr, "Error request pipe\n");
+    return 1;
+  }
 
   //TODO: close pipes
   if (unlink(client->resp_pipe_path) != 0 && errno != ENOENT) {
@@ -111,8 +128,14 @@ int ems_create(unsigned int event_id, size_t num_rows, size_t num_cols) {
     fprintf(stderr, "[ERR]: open server pipe failed: %s\n", strerror(errno));
     return 1;
   }
-  print_str_size(req_fd, message, 1 + sizeof(unsigned int) + 2 * sizeof(size_t));
-  close(req_fd);
+  if (print_str_size(req_fd, message, 1 + sizeof(unsigned int) + 2 * sizeof(size_t))) {
+    fprintf(stderr, "Error writing to request pipe\n");
+    return 1;
+  }
+  if(close(req_fd) == -1) {
+    fprintf(stderr, "Error request pipe\n");
+    return 1;
+  }
 
   // Open response pipe to receive
   int resp_fd = open(client->resp_pipe_path, O_RDONLY);
@@ -126,10 +149,16 @@ int ems_create(unsigned int event_id, size_t num_rows, size_t num_cols) {
   ssize_t bytesRead = read(resp_fd, response, sizeof(response));
   if (bytesRead == -1) {
     fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
-    close(resp_fd);
+    if (close(resp_fd) == -1) {
+      fprintf(stderr, "Error response pipe\n");
+      return 1;
+    }
     return 1;
   }
-  close(resp_fd);
+  if (close(resp_fd) == -1) {
+    fprintf(stderr, "Error response pipe\n");
+    return 1;
+  }
 
   int response_val;
   memcpy(&response_val, &response, sizeof(int));
@@ -153,8 +182,14 @@ int ems_reserve(unsigned int event_id, size_t num_seats, size_t* xs, size_t* ys)
     fprintf(stderr, "[ERR]: open server pipe failed: %s\n", strerror(errno));
     return 1;
   }
-  print_str_size(req_fd, message, 1 + sizeof(unsigned int) + sizeof(size_t) + 2 * (num_seats * sizeof(size_t)));
-  close(req_fd);
+  if (print_str_size(req_fd, message, 1 + sizeof(unsigned int) + sizeof(size_t) + 2 * (num_seats * sizeof(size_t)))) {
+    fprintf(stderr, "Error writing to request pipe\n");
+    return 1;
+  }
+  if(close(req_fd) == -1) {
+    fprintf(stderr, "Error request pipe\n");
+    return 1;
+  }
 
   // Open response pipe to receive
   int resp_fd = open(client->resp_pipe_path, O_RDONLY);
@@ -168,10 +203,16 @@ int ems_reserve(unsigned int event_id, size_t num_seats, size_t* xs, size_t* ys)
   ssize_t bytesRead = read(resp_fd, response, sizeof(response));
   if (bytesRead == -1) {
     fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
-    close(resp_fd);
+    if(close(resp_fd) == -1) {
+      fprintf(stderr, "Error response pipe\n");
+      return 1;
+    }
     return 1;
   }
-  close(resp_fd);
+  if (close(resp_fd) == -1) {
+    fprintf(stderr, "Error response pipe\n");
+    return 1;
+  }
 
   int response_val;
   memcpy(&response_val, &response, sizeof(int));
@@ -192,8 +233,14 @@ int ems_show(int out_fd, unsigned int event_id) {
     fprintf(stderr, "[ERR]: open request pipe failed: %s\n", strerror(errno));
     return 1;
   }
-  print_str_size(req_fd, message, 1 + sizeof(unsigned int));
-  close(req_fd);
+  if (print_str_size(req_fd, message, 1 + sizeof(unsigned int))) {
+    fprintf(stderr, "Error writing to request pipe\n");
+    return 1;
+  }
+  if (close(req_fd) == -1) {
+    fprintf(stderr, "Error request pipe\n");
+    return 1;
+  }
 
   // Open response pipe to receive
   int resp_fd = open(client->resp_pipe_path, O_RDONLY);
@@ -204,13 +251,19 @@ int ems_show(int out_fd, unsigned int event_id) {
 
   // Read from pipe
   char *response = malloc(sizeof(int) + 2 * sizeof(size_t));
+  if (response == NULL) {
+    fprintf(stderr, "Error allocating memory\n");
+    return 1;
+  }
   ssize_t bytesRead = read(resp_fd, response, sizeof(int) + 2 * sizeof(size_t));
   if (bytesRead == -1) {
     fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
-    close(resp_fd);
+    if(close(resp_fd) == -1) {
+      fprintf(stderr, "Error reponse pipe\n");
+      return 1;
+    }
     return 1;
   }
-  //close(resp_fd);
   int response_val;
   size_t num_rows, num_cols;
   memcpy(&response_val, response, sizeof(int));
@@ -224,15 +277,25 @@ int ems_show(int out_fd, unsigned int event_id) {
 
   free(response);
   response = malloc(num_rows * num_cols * sizeof(size_t));
-
-  size_t data_index = 0;
-  bytesRead = read(resp_fd, response, (num_rows * num_cols) * sizeof(size_t));
-  if (bytesRead == -1) {
-    fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
-    close(resp_fd);
+  if(response == NULL) {
+    fprintf(stderr, "Error allocating memory\n");
     return 1;
   }
-  close(resp_fd);
+
+  
+  bytesRead = read(resp_fd, response, (num_rows * num_cols) * sizeof(unsigned int));
+  if (bytesRead == -1) {
+    fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
+    if(close(resp_fd) == -1) {
+      fprintf(stderr, "Error response pipe\n");
+      return 1;
+    }
+    return 1;
+  }
+  if (close(resp_fd) == -1) {
+    fprintf(stderr, "Error response pipe\n");
+    return 1;
+  }
 
   // Write to output file
   for (size_t i = 1; i <= num_rows; i++) {
@@ -240,7 +303,7 @@ int ems_show(int out_fd, unsigned int event_id) {
       char buffer[16];
       size_t a;
       memcpy(&a, response + (sizeof(size_t) * ((i - 1) * num_cols + j - 1)), sizeof(size_t));
-      sprintf(buffer, "%u", a);
+      sprintf(buffer, "%lu", a);
 
       if (print_str(out_fd, buffer)) {
         perror("Error writing to file descriptor");
@@ -276,8 +339,14 @@ int ems_list_events(int out_fd) {
     fprintf(stderr, "[ERR]: open request pipe failed: %s\n", strerror(errno));
     return 1;
   }
-  print_str_size(req_fd, message, 1);
-  close(req_fd);
+  if (print_str_size(req_fd, message, 1)) {
+    fprintf(stderr, "Error writing to request pipe\n");
+    return 1;
+  };
+  if(close(req_fd) == -1) {
+    fprintf(stderr, "Error request pipe\n");
+    return 1;
+  }
 
   // Open response pipe to receive
   int resp_fd = open(client->resp_pipe_path, O_RDONLY);
@@ -288,10 +357,17 @@ int ems_list_events(int out_fd) {
 
   // Read from pipe
   char *response = malloc(sizeof(int) + sizeof(size_t));
+  if (response == NULL) {
+    fprintf(stderr, "Error allocating memory\n");
+    return 1;
+  }
   ssize_t bytesRead = read(resp_fd, response, sizeof(int) + sizeof(size_t));
   if (bytesRead == -1) {
     fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
-    close(resp_fd);
+    if(close(resp_fd) == -1) {
+      fprintf(stderr, "Error response pipe\n");
+      return 1;
+    }
     return 1;
   }
   int response_val;
@@ -305,31 +381,48 @@ int ems_list_events(int out_fd) {
 
   free(response);
   response = malloc(num_events * sizeof(unsigned int));
+  if (response == NULL) {
+    fprintf(stderr, "Error allocating memory\n");
+    return 1;
+  }
 
-  size_t data_index = 0;
   bytesRead = read(resp_fd, response, num_events * sizeof(unsigned int));
   if (bytesRead == -1) {
     fprintf(stderr, "[ERR]: read from response pipe failed: %s\n", strerror(errno));
-    close(resp_fd);
+    if (close(resp_fd) == -1) {
+      fprintf(stderr, "Error response pipe\n");
+      return 1;
+    }
     return 1;
   }
-  close(resp_fd);
+  if(close(resp_fd) == -1) {
+    fprintf(stderr, "Error response pipe\n");
+    return 1;
+  }
 
   // Write to output file
   if (num_events == 0) {
-    print_str(out_fd, "No events\n");
+    if (print_str(out_fd, "No events\n")) {
+      fprintf(stderr, "Error writing to stdout\n");
+      return 1;
+    }
   }
   else {
     for(size_t i = 0; i < num_events; i++) {
-      print_str(out_fd, "Event ");
+      if (print_str(out_fd, "Event ")) {
+        fprintf(stderr, "Error writing to stdout\n");
+        return 1;
+      }
       unsigned int temp;
       memcpy(&temp, response + i * sizeof(unsigned int), sizeof(unsigned int));
 
       char id[16];
-      sprintf(id, "%u\n", temp, sizeof(unsigned int));
-      print_str(out_fd, id);
+      sprintf(id, "%u\n", temp);
+      if(print_str(out_fd, id)) {
+        fprintf(stderr, "Error writing to stdout\n");
+        return 1;
+      }
     }
   }
-
   return response_val;
 }
